match run_process name to service name
The run_process name has to be the same as the enable_service name
otherwise run_process won't actually run that thing.

The service name was changed in this plugin to 0mq to ensure the
plugin fully stands on it's own, and no vestigial pieces of the old
code is needed for it to work.