import zuul job settings from project-config
This is a mostly mechanically generated patch to complete step 1 of
moving the zuul job settings out of project-config and into each
project repository.

Because there will be a separate patch on each branch, the branch
specifiers for branch-specific jobs have been removed.

The legacy-tempest-dsvm-full-zmq job is only used by this plugin, so
we move the playbook and job definition here as well. It needs to be
renamed in order to be moved and applied to the repository, so the new
name devstack-plugin-zmq-tempest-full is used to indicate that it is
defined within the repository.

See the python3-first goal document for details:
https://governance.openstack.org/tc/goals/stein/python3-first.html

Change-Id: I3d80c8a576fa4cbedffdaaf6cf97a4a61a2dce34
Story: #2003250